DEF-3813 - Bob should now show better resource information for certain errors. (#4171)

+++ com.dynamo.cr/com.dynamo.cr.bob/src/com/dynamo/bob/bundle/BundleHelper.java
@@ -76,7 +76,7 @@ public class BundleHelper {
         }
     }
 
-    public BundleHelper(Project project, Platform platform, File bundleDir, String appDirSuffix, String variant) throws IOException {
+    public BundleHelper(Project project, Platform platform, File bundleDir, String appDirSuffix, String variant) throws CompileExceptionError {
         this.projectProperties = project.getProjectProperties();
 
         this.project = project;
@@ -86,7 +86,11 @@ public class BundleHelper {
         this.buildDir = new File(project.getRootDirectory(), project.getBuildDirectory());
         this.appDir = new File(bundleDir, title + appDirSuffix);
 
-        this.propertiesMap = createPropertiesMap(project.getProjectProperties());
+        try {
+            this.propertiesMap = createPropertiesMap(project.getProjectProperties());
+        } catch (IOException e) {
+            throw new CompileExceptionError(project.getGameProjectResource(), -1, e);
+        }
 
         this.variant = variant;
     }
@@ -182,6 +186,9 @@ public class BundleHelper {
     private List<File> formatAll(List<IResource> sources, File toDir, Map<String, Object> properties) throws IOException {
         List<File> out = new ArrayList<File>();
         for (IResource source : sources) {
+            if (!source.exists()) {
+                throw new IOException(String.format("Resource %s does not exist.", source));
+            }
             // converts a relative path to something we can easily see if we get a merge failure: a/b/c.xml -> a_b_c.xml
             String name = source.getPath().replaceAll("[^a-zA-Z0-9_]", "_");
             File target = new File(toDir, name);
@@ -202,7 +209,7 @@ public class BundleHelper {
         } else if (platform == Platform.JsWeb || platform == Platform.WasmWeb) {
             name = BundleHelper.MANIFEST_NAME_HTML5;
         } else {
-            throw new CompileExceptionError(null, -1, "Unsupported ManifestMergeTool platform: " + platform.toString());
+            throw new CompileExceptionError(mainManifest, -1, "Unsupported ManifestMergeTool platform: " + platform.toString());
         }
 
         // First, list all manifests
@@ -233,14 +240,14 @@ public class BundleHelper {
         } else if (platform == Platform.JsWeb || platform == Platform.WasmWeb) {
             manifestPlatform = ManifestMergeTool.Platform.HTML5;
         } else {
-            throw new CompileExceptionError(null, -1, "Merging manifests for platform unsupported: " + platform.toString());
+            throw new CompileExceptionError(mainManifest, -1, "Merging manifests for platform unsupported: " + platform.toString());
         }
 
         // Now merge these manifests in order (the main manifest is first)
         try {
             ManifestMergeTool.merge(manifestPlatform, resolvedMainManifest, outManifest, resolvedManifests);
         } catch (RuntimeException e) {
-            throw new CompileExceptionError(null, -1, "Failed merging manifests: " + e.toString());
+            throw new CompileExceptionError(mainManifest, -1, "Failed merging manifests: " + e.toString());
         }
         FileUtils.deleteDirectory(manifestDir);
     }
@@ -379,11 +386,15 @@ public class BundleHelper {
         }
     }
 
-    public List<ExtenderResource> generateAndroidResources(Project project, File resDir, File manifestFile, File apk, File tmpDir) throws CompileExceptionError, IOException {
+    public List<ExtenderResource> generateAndroidResources(Project project, File resDir, File manifestFile, File apk, File tmpDir) throws CompileExceptionError {
 
         // Get all Android specific resources needed to create R.java files
-        BundleHelper.createAndroidResourceFolders(resDir);
-        copyAndroidIcons(resDir);
+        try {
+            BundleHelper.createAndroidResourceFolders(resDir);
+            copyAndroidIcons(resDir);
+        } catch (Exception e) {
+            throw new CompileExceptionError(project.getGameProjectResource(), -1, e);
+        }
 
         // We store the extensions' resources in a separate folder, because they otherwise failed on the Android naming convention.
         // I.e. resDir contains asset directories, extensionsDir contains package directories that contain asset directiores
@@ -466,15 +477,18 @@ public class BundleHelper {
         return resizedImage;
     }
 
-    private void genIcon(BufferedImage fallbackImage, File outputDir, String propertyName, String outName, int size) throws IOException
+    private void genIcon(BufferedImage fallbackImage, File outputDir, String propertyCategory, String propertyName, String outName, int size) throws IOException
     {
         File outFile = new File(outputDir, outName);
 
         // If the property was found just copy icon file to output folder.
         if (propertyName.length() > 0) {
-            String resource = projectProperties.getStringValue("ios", propertyName);
+            String resource = projectProperties.getStringValue(propertyCategory, propertyName);
             if (resource != null && resource.length() > 0) {
                 IResource inResource = project.getResource(resource);
+                if (!inResource.exists()) {
+                    throw new IOException(String.format("%s does not exist.", resource));
+                }
                 FileUtils.writeByteArrayToFile(outFile, inResource.getContent());
                 return;
             }
@@ -498,6 +512,9 @@ public class BundleHelper {
 
         if (largestIcon != null) {
             IResource largestIconRes = project.getResource(largestIcon);
+            if (!largestIconRes.exists()) {
+                throw new IOException("Could not find resource: " + largestIcon);
+            }
             FileUtils.writeByteArrayToFile(largestIconFile, largestIconRes.getContent());
         } else {
             URL defaultIconURL = getClass().getResource("resources/ios/default_icon.png");
@@ -514,15 +531,15 @@ public class BundleHelper {
         BufferedImage largestIconImage = getFallbackIconImage("ios", iconPropNames);
 
         // Copy game.project specified icons
-        genIcon(largestIconImage, appDir,   "app_icon_57x57",       "Icon.png",  57);
-        genIcon(largestIconImage, appDir, "app_icon_114x114",    "Icon@2x.png", 114);
-        genIcon(largestIconImage, appDir,   "app_icon_72x72",    "Icon-72.png",  72);
-        genIcon(largestIconImage, appDir, "app_icon_144x144", "Icon-72@2x.png", 144);
-        genIcon(largestIconImage, appDir,   "app_icon_76x76",    "Icon-76.png",  76);
-        genIcon(largestIconImage, appDir, "app_icon_152x152", "Icon-76@2x.png", 152);
-        genIcon(largestIconImage, appDir, "app_icon_120x120", "Icon-60@2x.png", 120);
-        genIcon(largestIconImage, appDir, "app_icon_180x180", "Icon-60@3x.png", 180);
-        genIcon(largestIconImage, appDir, "app_icon_167x167",   "Icon-167.png", 167);
+        genIcon(largestIconImage, appDir, "ios",   "app_icon_57x57",       "Icon.png",  57);
+        genIcon(largestIconImage, appDir, "ios", "app_icon_114x114",    "Icon@2x.png", 114);
+        genIcon(largestIconImage, appDir, "ios",   "app_icon_72x72",    "Icon-72.png",  72);
+        genIcon(largestIconImage, appDir, "ios", "app_icon_144x144", "Icon-72@2x.png", 144);
+        genIcon(largestIconImage, appDir, "ios",   "app_icon_76x76",    "Icon-76.png",  76);
+        genIcon(largestIconImage, appDir, "ios", "app_icon_152x152", "Icon-76@2x.png", 152);
+        genIcon(largestIconImage, appDir, "ios", "app_icon_120x120", "Icon-60@2x.png", 120);
+        genIcon(largestIconImage, appDir, "ios", "app_icon_180x180", "Icon-60@3x.png", 180);
+        genIcon(largestIconImage, appDir, "ios", "app_icon_167x167",   "Icon-167.png", 167);
     }
 
     public void copyAndroidIcons(File resDir) throws IOException
@@ -532,13 +549,13 @@ public class BundleHelper {
         BufferedImage largestIconImage = getFallbackIconImage("android", iconPropNames);
 
         // copy old 32x32 icon first, the correct size is actually 36x36
-        genIcon(largestIconImage, resDir,   "app_icon_32x32",    "drawable-ldpi/icon.png",  36);
-        genIcon(largestIconImage, resDir,   "app_icon_36x36",    "drawable-ldpi/icon.png",  36);
-        genIcon(largestIconImage, resDir,   "app_icon_48x48",    "drawable-mdpi/icon.png",  48);
-        genIcon(largestIconImage, resDir,   "app_icon_72x72",    "drawable-hdpi/icon.png",  72);
-        genIcon(largestIconImage, resDir,   "app_icon_96x96",   "drawable-xhdpi/icon.png",  96);
-        genIcon(largestIconImage, resDir, "app_icon_144x144",  "drawable-xxhdpi/icon.png", 144);
-        genIcon(largestIconImage, resDir, "app_icon_192x192", "drawable-xxxhdpi/icon.png", 192);
+        genIcon(largestIconImage, resDir, "android",   "app_icon_32x32",    "drawable-ldpi/icon.png",  36);
+        genIcon(largestIconImage, resDir, "android",   "app_icon_36x36",    "drawable-ldpi/icon.png",  36);
+        genIcon(largestIconImage, resDir, "android",   "app_icon_48x48",    "drawable-mdpi/icon.png",  48);
+        genIcon(largestIconImage, resDir, "android",   "app_icon_72x72",    "drawable-hdpi/icon.png",  72);
+        genIcon(largestIconImage, resDir, "android",   "app_icon_96x96",   "drawable-xhdpi/icon.png",  96);
+        genIcon(largestIconImage, resDir, "android", "app_icon_144x144",  "drawable-xxhdpi/icon.png", 144);
+        genIcon(largestIconImage, resDir, "android", "app_icon_192x192", "drawable-xxxhdpi/icon.png", 192);
     }
 
     private boolean copyIcon(BobProjectProperties projectProperties, String projectRoot, File resDir, String name, String outName)
